dom/tests/mochitest/pointerlock/file_unlockOnUnload.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Bug 633602 - file_unlockOnUnload.html</title>
  <script type="text/javascript"
          src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript"
          src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript"
          src="mouselock_util.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style type="text/css">
      #test-element { background-color: #94E01B; width:100px; height:100px; }
  </style>
</head>
<body onload="start();" onunload="runUnload();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
    Mozilla Bug 633602</a>
  <div id="content">
    <div id="test-element"></div>
  </div>
  <pre id="test">
    <script type="text/javascript">
      /*
       * Test for Bug 633602 
       * This test checks to ensure that mouselock is lost when window unloads
       */
      SimpleTest.waitForExplicitFinish();
      var pointer = navigator.mozPointer;
      var testElement = document.getElementById("test-element");

      document.addEventListener("mozfullscreenchange", function(e) {
        if(document.mozFullScreen &&
            document.mozFullScreenElement === testElement) {
          pointer.lock(testElement, function () {
              // After lock, close the window to check for "Unload"
              window.close();
            }, function () {
              ok(pointer.isLocked, "Mouse failed to lock fullscreen element");
              SimpleTest.finish();
          });
        }
      }, false);

      function start() {
        SimpleTest.waitForFocus(function() {
          testElement.mozRequestFullScreen();
        });
      }

      function runUnload() {
        is(pointer.isLocked, false, "Mouse unlocked on unload");
        SimpleTest.finish();
      }
    </script>
  </pre>
</body>
</html>
